invoker/pyproject.toml
[tool.poetry]
name = "invoker"
version = "0.1.0"
description = ""
authors = ["Prince Baghel <pkbghl90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dapr = "^1.14.0"
dapr-ext-grpc = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
